Refactor some utils which might be usable elsewhere, not just in the prepare script
