+++ stanza/utils/datasets/ner/prepare_ner_dataset.py
@@ -257,6 +257,7 @@ import stanza.utils.datasets.ner.convert_starlang_ner as convert_starlang_ner
 import stanza.utils.datasets.ner.prepare_ner_file as prepare_ner_file
 import stanza.utils.datasets.ner.suc_to_iob as suc_to_iob
 import stanza.utils.datasets.ner.suc_conll_to_iob as suc_conll_to_iob
+from stanza.utils.datasets.ner.utils import convert_bio_to_json, get_tags, read_tsv, write_dataset
 
 SHARDS = ('train', 'dev', 'test')
 
@@ -264,99 +265,6 @@ class UnknownDatasetError(ValueError):
     def __init__(self, dataset, text):
         super().__init__(text)
         self.dataset = dataset
-
-def convert_bio_to_json(base_input_path, base_output_path, short_name, suffix="bio", shard_names=SHARDS):
-    """
-    Convert BIO files to json
-
-    It can often be convenient to put the intermediate BIO files in
-    the same directory as the output files, in which case you can pass
-    in same path for both base_input_path and base_output_path.
-    """
-    for input_shard, output_shard in zip(shard_names, SHARDS):
-        input_filename = os.path.join(base_input_path, '%s.%s.%s' % (short_name, input_shard, suffix))
-        if not os.path.exists(input_filename):
-            alt_filename = os.path.join(base_input_path, '%s.%s' % (input_shard, suffix))
-            if os.path.exists(alt_filename):
-                input_filename = alt_filename
-            else:
-                raise FileNotFoundError('Cannot find %s component of %s in %s or %s' % (output_shard, short_name, input_filename, alt_filename))
-        output_filename = os.path.join(base_output_path, '%s.%s.json' % (short_name, output_shard))
-        print("Converting %s to %s" % (input_filename, output_filename))
-        prepare_ner_file.process_dataset(input_filename, output_filename)
-
-def get_tags(datasets):
-    """
-    return the set of tags used in these datasets
-
-    datasets is expected to be train, dev, test but could be any list
-    """
-    tags = set()
-    for dataset in datasets:
-        for sentence in dataset:
-            for word, tag in sentence:
-                tags.add(tag)
-    return tags
-
-def write_dataset(datasets, output_dir, short_name, suffix="bio"):
-    """
-    write all three pieces of a dataset to output_dir
-
-    datasets should be 3 lists: train, dev, test
-    each list should be a list of sentences
-    each sentence is a list of pairs: word, tag
-
-    after writing to .bio files, the files will be converted to .json
-    """
-    for shard, dataset in zip(SHARDS, datasets):
-        output_filename = os.path.join(output_dir, "%s.%s.%s" % (short_name, shard, suffix))
-        with open(output_filename, "w", encoding="utf-8") as fout:
-            for sentence in dataset:
-                for word in sentence:
-                    fout.write("%s\t%s\n" % word)
-                fout.write("\n")
-
-    convert_bio_to_json(output_dir, output_dir, short_name, suffix)
-
-
-def read_tsv(filename, text_column, annotation_column, remap_fn=None, skip_comments=True):
-    """
-    Read sentences from a TSV file
-
-    Returns a list of list of (word, tag)
-    """
-    with open(filename, encoding="utf-8") as fin:
-        lines = fin.readlines()
-
-    lines = [x.strip() for x in lines]
-
-    sentences = []
-    current_sentence = []
-    for line in lines:
-        if not line:
-            if current_sentence:
-                sentences.append(current_sentence)
-                current_sentence = []
-            continue
-        if skip_comments and line.startswith("#"):
-            continue
-
-        pieces = line.split("\t")
-        word = pieces[text_column]
-        if word == '\x96':
-            # this happens in GermEval2014 for some reason
-            continue
-        tag = pieces[annotation_column]
-        if remap_fn:
-            tag = remap_fn(tag)
-
-        current_sentence.append((word, tag))
-
-    if current_sentence:
-        sentences.append(current_sentence)
-
-    return sentences
-
 
 def process_turku(paths, short_name):
     assert short_name == 'fi_turku'
